The revamp of the parser is done.
More faithful to Bash. More efficient. Relatively more straightforward
than the previous version.

The command line parsing is now composed of a parser and an expander.
The parser is also used by the completer.

When defining an alias, the value is always fully expanded to the form
where it contains no more alias. This removes the need of recursive alias
expanding when parsing command line.

Added primitive history search functions.