Background jobs now send output to builtin console instead of stash terminal.
Add httpserver.py script for simple http server with upload function.
This script is meant to run as a background job so that it can be
terminated by the close button on interactive prompt page. This allows
the port binding to be released. So the same command can be called again
without having to restart Pythonista.

The stdout redirect now also works for stderr.