Shell at each level now has three sets of variables belong to three types of scopes instead of two.
The three scopes are enclosed, current and enclosing. They are needed to
correctly emulate the sub-shell environment.

The first member in state stack is no longer special except its
environment is persisted back to its parent's environment. The top level
environment, i.e. parent of the first stack member is made clear that it
is not stored in the stack and has empty enclosed scope (as it is the
top). It provides the persisting environment for the first stack member
to use.

load_rcfile now uses the StaSh __call__() interface.

Pure comment lines are now allowed in shell scripts.