Streamlined many of the i/o methods.
write now has a rng keyword. This is different from file-like object.
Since we are having a free-navigable screen, it makes sense to have a
range specified.

Made it clear that generating a new input prompt is not the same as set
the text to an existing one. Prompt position is always recorded so input
read is more accurate.