Fixed various bugs in the new completion with cursor position.
Sub-cmd completion now takes a single parameter that is the list of
tokens with the last one being the word_to_complete.
Completion at cursor now retain the cursor position after completion.
If screen buffer is being halved, do not bother with cursor_rindex as it
sometimes make cursor disappear.