Added ctypes thread with async raise method.
This is an alternative to the traced thread as a killable thread.
The thread type can be configured via the config file. Currently, ctypes
thread is the default choice due to much better performance.
The biggest advantage of ctypes thread is the speed. It is about 3-4 times
faster than the traced threads. However the kill method is not as reliable
as the traced thread. Multiple kill call may be needed to actually stop a
thread and it is not feasible to automate it. So currently it is up to
users to press the C-C button multiple times when necessary.