Working commit.
os and sys can be imported and cached inside other modules, e.g. argparse.
This leads to argparse always sees the same sys.argv because it has its
own sys reference and not lookup for the mocked one.
It seems possible to force reload all modules for running script so that
all modules will have access to the mocked modules.
There are still many issues. Now I am not sure if it is worthwhile.