io.open is now renamed to io_open instead of open.
This is to avoid shadowing Python 2 builtin open. A major difference of
these two functions is that io.open returns unicode. This may not be
wanted for Python 2 users and packages are not fully unicode compliant.
An example is the man script that read a file and pass the content to
compile. Some file contains leading encoding declaration and this causes a
problem to compile when the file is read as an unicode string.