man.py: now supports help for non-commands
The `man.py` command now supports help for non-command.
`man XYZ` will first look wether it can find the command `XYZ`. If it does, it will display the docstring of `XYZ`.
If it cant find XYZ, it will try to find a file named `XYZ.*` in `$STASH_ROOT/man`. If it can find a file, it will display it depending on the extension.
If it can find a dir named XYZ in `$STASH_ROOT/man`, it will look a file named `page_1.*` inside it and display it.
`man XYZ(2)` will display a file named `page_2.*` in `$STASH_ROOT/man/XYZ/` if `$STASH_ROOT/man/XYZ` is a dir. Otherwise, it will ignore the pagenumber.
Currently supported extensions:
   - `.txt`: print the content to stdout
   - `.url`: open the content (without `\n` in `webviewer`)
   - `.html`: open the path in `quicklook`
   - all other extensions will be handled like `.txt`