added completion and documentation
I wrote some documentation about the monkeypatching.
To view it, type `man monkeypatching` in StaSh.
I have also updated libcompleter to contain auto-completion for
monkeylord.
I also have tried to implement a thread-local `sys.argv`-system but
failed (because it is overwritten by stash). The Test is avialable in
the INSTABLE Group.