Added extension support for FSIs
You can now define FSIs (=The classes used by 'mount' and 'mc' for
managing filesystems) as a extension. I have updated the 'man'-pages to
contain some instructions about this.The extensions are loaded the first
time 'stashutils.fsi.interfaces' is imported and may overwrite old
definitions with the same name.