Adding sed.py and according unittests
sed.py is a copy of sed.py from the github project fschaeck/PythonSed that I am the maintainer off.
It is a fully compatible Python implementation of GNU sed 4.8 and thus the GNU sed 4.8 documentation
can be used as a reference. See https://github.com/fschaeck/PythonSed for more information.